refactor(#3479): shrink completion_guard.rs below giant threshold
Verbatim-extract two cohesive leaf child modules out of
turn_bridge/completion_guard.rs (1834 -> 872 prod LoC), graduating it from
the giant-file registry:

- completion_guard/completion_postgres.rs (530 prod + tests): runtime-Postgres
  last-resort dispatch completion/failure paths, auto-queue reconciliation, and
  dispatch-followup / reconcile-marker plumbing.
- completion_guard/completion_context.rs (462 prod): work-dispatch completion
  context + commit attribution (hints lookup, git-history resolution,
  build_work_dispatch_completion_result, noop/tracked-change builders).

Behaviour-preserving: bodies are byte-identical except visibility widening
(pub(super) / preserved pub(in discord)/pub(crate)), one relocated super:: path
depth, and re-exports from completion_guard.rs so turn_bridge/mod.rs and all
external callers (recovery_engine, streaming_finalizer) resolve unchanged.

Giant-file graduation: removed completion_guard.rs from giant_file_registry
grandfathered + grandfathered_baseline_paths (win-locked) and retired its
ratchet baseline; regenerated ARCHITECTURE.md and synced change-surfaces.md.

Gates: cargo check (0 errors), cargo test --lib completion_guard (3 pass),
audit_maintainability.py --check (rc 0), generate_inventory_docs.py (rc 0),
ci-script-checks.sh (rc 0).
