+++ scripts/audit_maintainability_giant_baseline.toml
@@ -304,8 +304,8 @@
 # suppressed-label noise, user report 2026-06-12: +38 for provider-aware
 # status/footer stripping in the placeholder suppression decisions.
 "src/services/discord/tmux.rs" = 2049
-# #3034 batch-8: lowered 1849 -> 1834 (deleted dead `build_verdict_payload`).
-"src/services/discord/turn_bridge/completion_guard.rs" = 1834
+# completion_guard.rs decomposed in #3479 (1834 -> 872 prod LoC); dropped below
+# the giant threshold, so its ratchet baseline is retired (no longer tracked).
 # #3034 batch-8: lowered 1781 -> 1753 (deleted the dead `assistant_turns` turn-cap
 # tracker — field + 3 methods + construction sites).
 "src/services/discord/session_runtime.rs" = 1753
